+++ src/mvalue/mvalue.jl
@@ -1,6 +1,6 @@
 using PDBTools
 
-export mvalue, parse_mvalue_server_sasa, gmx_sasa
+export mvalue, parse_mvalue_server_sasa, gmx_delta_sasa_per_restype, delta_sasa_per_restype
 export MoeserHorinek, AutonBolen
 #export sasa_desnat_average
 
@@ -11,7 +11,7 @@ struct AutonBolen <: MvalueModel end
 include("./data.jl")
 
 """
-    mvalue(; model=MoeserHorinek, cosolvent="urea", pdbname, sasas, type=1)
+    mvalue(; model=MoeserHorinek, cosolvent="urea", atoms:AbstractVector{<:PDBTools.Atom}, sasas, type=1)
 
 Calculates the m-value (transfer free energy of a protein in 1M solution) using the Tanford transfer model,
 as implemented by Moeser and Horinek [1] or by Auton and Bolen [2,3].
@@ -21,11 +21,12 @@ as implemented by Moeser and Horinek [1] or by Auton and Bolen [2,3].
 - `model`: The model to be used. Must be `MoeserHorinek` or `AutonBolen`. `MoeserHorinek` is only implemented for `cosolvent="urea"`,
    and should be more precise in that case. Other solvents are available for `AutonBolen`.
 - `cosolvent::String`: One of $(join('"' .* sort!(unique(keys(MolSimToolkit.cosolvent_column)) .* '"'; by=lowercase),", "))
-- `pdbname::AbstractString`: Path to the PDB file of the protein structure.
+- `atoms::AbstractVector{<:PDBTools.Atom}`: Vector containing the atoms of the structure.
 - `sasas::Dict{String, Dict{Symbol, Float64}}`: A dictionary containing the change in solvent accessible surface area (SASA)
   upon denaturation for each amino acid type. This data can be obtained from the m-value server or calculated using GROMACS:
     - The output of the server can be parsed using the `parse_mvalue_server_sasa` function defined in this module.
-    - Alternatively, the SASA values can be calculated using GROMACS with the `gmx_sasa` function defined in this module.
+    - Compute the SASA with `delta_sasa_per_restype`, a SASA calculation utility implemented in PDBTools.jl.
+    - SASA values can be calculated using GROMACS with the `gmx_delta_sasa_per_restype` function defined in this module.
 
 - `type::Int`: Specifies which SASA value to use from the provided data, because the server provides minimum, average, and maximum values,
     according to different denatured models for the protein. The recommended value is `2` for comparison with experimental data.
@@ -44,13 +45,20 @@ Each entry in the dictionary is a named tuple with `bb` and `sc` fields represen
 # Example calls
 
 ```julia
+using MolSimToolkit
+protein = read_pdb("protein.pdb")
+
+# Using SASA values calculated with PDBTools.jl
+sasas=delta_sasa_per_restype(native=read_pdb("native.pdb"), desnat=read_pdb("desnat.pdb"))
+mvalue(; model=AutonBolen, cosolvent="TMAO", atoms=protein, sasas=sasas)
+
 # Using SASA values from the m-value server
 sasas_from_server=parse_mvalue_server_sasa(server_output)
-mvalue(; model=MoeserHorinek, cosolvent="urea", pdbname="protein.pdb", sasas=sasas_from_server, type=2)
+mvalue(; model=MoeserHorinek, cosolvent="urea", atoms=protein, sasas=sasas_from_server, type=2)
 
 # Using SASA values calculated with GROMACS
-sasas_gmx=gmx_sasa(native_pdb="native.pdb", desnat_pdb="desnat.pdb")
-mvalue(; model=AutonBolen, cosolvent="TMAO", pdbname="protein.pdb", sasas=sasas_gmx, type=1)
+sasas_gmx=gmx_delta_sasa_per_restype(native_pdb="native.pdb", desnat_pdb="desnat.pdb")
+mvalue(; model=AutonBolen, cosolvent="TMAO", atoms=protein, sasas=sasas_gmx)
 ```
 
 ## References
@@ -63,14 +71,15 @@ mvalue(; model=AutonBolen, cosolvent="TMAO", pdbname="protein.pdb", sasas=sasas_
 function mvalue(;
     model::Type{<:MvalueModel}=MoeserHorinek,
     cosolvent::String="urea",
-    pdbname, sasas, type=1
+    atoms::AbstractVector{<:PDBTools.Atom}, sasas, type=1
 )
-    protein = read_pdb(pdbname, "protein")
-    residue_types = unique(threeletter.(resname.(protein)))
+    protein = select(atoms, "protein")
+    residue_names = unique(resname.(protein))
     DeltaG_per_residue = Dict{String,@NamedTuple{bb::Float64, sc::Float64}}()
-    for rname in residue_types
-        DeltaG_per_residue[rname] = (bb=(last(tfe_asa(model, cosolvent, rname))) * (sasas[rname][:bb][type] / 100),
-            sc=(first(tfe_asa(model, cosolvent, rname))) * (sasas[rname][:sc][type] / 100))
+    for rname in residue_names
+        rtype = threeletter(rname) # convert non-standard residue names in types (e. g. HSD -> HIS)
+        DeltaG_per_residue[rtype] = (bb=(last(tfe_asa(model, cosolvent, rtype))) * (sasas[rname][:bb][type] / 100),
+            sc=(first(tfe_asa(model, cosolvent, rtype))) * (sasas[rname][:sc][type] / 100))
     end
     DeltaG_BB = sum(getfield(DeltaG_per_residue[key], :bb) for key in keys(DeltaG_per_residue))
     DeltaG_SC = sum(getfield(DeltaG_per_residue[key], :sc) for key in keys(DeltaG_per_residue))
@@ -114,24 +123,64 @@ function tfe_asa(
     return sc_contribution / 10, bb_contribution / 10
 end
 
-#=
-    read_gmx_sasa_values(filename::String, n)
+"""
+    delta_sasa_per_restype(; 
+        native::AbstractVector{<:PDBTools.Atom}, 
+        desnat::AbstractVector{<:PDBTools.Atom}
+    )
 
-Reads the output of `gmx sasa` and returns the SASA values.
-`n` is the number of surfaces calculated (1 for BB only, 2 for SC and BB, for example).
+Calculates the change in solvent accessible surface area (SASA) upon denaturation for each amino acid type
+using PDBTools. Returns a dictionary that can be directly used as input to the `mvalue` function.
 
-=#
-function read_gmx_sasa_values(filename::String, n)
-    local sasa_values
-    open(filename, "r") do io
-        for line in eachline(io)
-            if startswith(line, r"@|#")
-                continue
-            end
-            sasa_values = parse.(Float64, split(line)[3:2+n])
+# Arguments
+
+- `native`: Vector of PDBTools.Atom objects for the native structure.
+- `desnat`: Vector of PDBTools.Atom objects for the denatured structure.
+
+# Returns
+
+A dictionary where each key is an amino acid three-letter code (e.g., "ALA", "PHE"), and the value
+is another dictionary with two keys: `:sc` for side chain SASA values and `:bb` for backbone SASA values.
+Each of these keys maps to a tuple containing a single Float64 value representing the change in SASA upon denaturation in Å².
+
+# Optional arguments
+
+- `n_dots::Int=500`: Sets the precision of the SASA calculation (greater is better).
+- `backbone::Function = at -> name(at) in ("N", "CA", "C", "O")`: Define what is a backbone atom.
+- `sidechain::Function = at -> !(name(at) in ("N", "CA", "C", "O"))`: Define what is a sidechain atom.
+- `ignore_hydrogen::Bool = true`: By default, ignore all Hydrogen atoms of the structure.
+- `unitcell=nothing`: By default, do not use periodic boundary conditions. To use PBCs, define A
+  unitcell by providing either a 3x3 matrix or, for orthorhombic cells, a vector of length 3 of cell sides.
+
+"""
+function delta_sasa_per_restype(;
+    native::AbstractVector{<:PDBTools.Atom},
+    desnat::AbstractVector{<:PDBTools.Atom},
+    n_dots::Int=500,
+    backbone::Function=at -> name(at) in ("N", "CA", "C", "O"),
+    sidechain::Function=at -> !(name(at) in ("N", "CA", "C", "O")),
+    ignore_hydrogen::Bool=true,
+    unitcell=nothing,
+)
+    keepH(at) = ignore_hydrogen ? !(element(at) == "H") : true
+    native_atoms = select(native, at -> isprotein(at) & keepH(at))
+    desnat_atoms = select(desnat, at -> isprotein(at) & keepH(at))
+    native_atomic_sasa = PDBTools.sasa_particles(native_atoms; n_dots, unitcell)
+    desnat_atomic_sasa = PDBTools.sasa_particles(desnat_atoms; n_dots, unitcell)
+    sasas = Dict{String,Dict{Symbol,Float32}}()
+    for rname in unique(resname.(eachresidue(native_atoms)))
+        bb_native = PDBTools.sasa(native_atomic_sasa, at -> resname(at) == rname && backbone(at))
+        bb_desnat = PDBTools.sasa(desnat_atomic_sasa, at -> resname(at) == rname && backbone(at))
+        if !ignore_hydrogen || rname != "GLY"
+            sc_native = PDBTools.sasa(native_atomic_sasa, at -> resname(at) == rname && sidechain(at))
+            sc_desnat = PDBTools.sasa(desnat_atomic_sasa, at -> resname(at) == rname && sidechain(at))
+        else
+            sc_native = 0.0
+            sc_desnat = 0.0
         end
+        sasas[rname] = Dict(:sc => sc_desnat - sc_native, :bb => bb_desnat - bb_native)
     end
-    return sasa_values
+    return sasas # Å^2  
 end
 
 """
@@ -183,19 +232,48 @@ function parse_mvalue_server_sasa(string::AbstractString)
     return sasa
 end
 
+#=
+    read_gmx_delta_sasa_per_restype_values(filename::String, n)
+
+Reads the output of `gmx sasa` and returns the SASA values.
+`n` is the number of surfaces calculated (1 for BB only, 2 for SC and BB, for example).
+
+=#
+function read_gmx_delta_sasa_per_restype_values(filename::String, n)
+    local sasa_values
+    open(filename, "r") do io
+        for line in eachline(io)
+            if startswith(line, r"@|#")
+                continue
+            end
+            sasa_values = parse.(Float64, split(line)[3:2+n])
+        end
+    end
+    return sasa_values
+end
+
 """
-    gmx_sasa(; native_pdb::AbstractString, desnat_pdb::AbstractString)
+    gmx_delta_sasa_per_restype(; native_pdb::AbstractString, desnat_pdb::AbstractString)
 
 Calculates the change in solvent accessible surface area (SASA) upon denaturation for each amino acid type
 using GROMACS. Returns a dictionary that can be directly used as input to the `mvalue` function.
 
 !!! note
-    This function requires GROMACS (`gmx sasa` executable) to be installed and accessible from the command line. 
+    This function requires GROMACS (`gmx sasa` executable) to be installed and accessible from the command line.
+    The path to the `gmx` executable can be provided with the `gmx` keyword.
 
 # Arguments
 
 - `native_pdb::AbstractString`: Path to the PDB file of the native protein structure.
 - `desnat_pdb::AbstractString`: Path to the PDB file of the denatured protein structure.
+
+# Optional arguments 
+
+- `gmx`: the path to the `gmx` GROMACS exectuable (by default it expects `gmx` to be on the path).
+- `n_dots::Int`: sets the precision of the SASA grid (greater is better).
+- `backbone::Function = at -> name(at) in ("N", "CA", "C", "O")`: Define what is a backbone atom.
+- `sidechain::Function = at -> !(name(at) in ("N", "CA", "C", "O"))`: Define what is a sidechain atom.
+- `ignore_hydrogen::Bool=true`: By default, ignore all hydrogen atoms.
 
 # Returns
 
@@ -204,15 +282,20 @@ is another dictionary with two keys: `:sc` for side chain SASA values and `:bb` 
 Each of these keys maps to a tuple containing a single Float64 value representing the change in SASA upon denaturation in Å².
 
 """
-function gmx_sasa(;
+function gmx_delta_sasa_per_restype(;
     native_pdb::AbstractString,
     desnat_pdb::AbstractString,
+    n_dots::Int=500,
+    ignore_hydrogen::Bool=true,
+    gmx="gmx",
+    backbone::Function=at -> name(at) in ("N", "CA", "C", "O"),
+    sidechain::Function=at -> !(name(at) in ("N", "CA", "C", "O")),
 )
     sasas = Dict{String,Dict{Symbol,Float64}}()
     p = read_pdb(native_pdb, "protein")
     for rname in unique(resname.(eachresidue(p)))
-        sasa_bb_native, sasa_sc_native = 100 .* gmx_sasa_single(native_pdb, rname) # returns in nm^2
-        sasa_bb_desnat, sasa_sc_desnat = 100 .* gmx_sasa_single(desnat_pdb, rname)
+        sasa_bb_native, sasa_sc_native = 100 .* gmx_sasa_per_restype(native_pdb, rname; gmx, n_dots, backbone, sidechain, ignore_hydrogen) # returns in nm^2
+        sasa_bb_desnat, sasa_sc_desnat = 100 .* gmx_sasa_per_restype(desnat_pdb, rname; gmx, n_dots, backbone, sidechain, ignore_hydrogen)
         sasas[rname] = Dict(:sc => sasa_sc_desnat - sasa_sc_native, :bb => sasa_bb_desnat - sasa_bb_native)
     end
     return sasas # returns in Å^2
@@ -221,13 +304,28 @@ end
 #
 # Runs gmx sasa for a single residue type in a given PDB file.
 #
-function gmx_sasa_single(pdbname, resname)
+function gmx_sasa_per_restype(
+    pdbname, rname;
+    gmx="gmx",
+    n_dots::Int=500,
+    backbone::Function,
+    sidechain::Function,
+    ignore_hydrogen::Bool=true,
+)
+    keepH(at) = ignore_hydrogen ? !(element(at) == "H") : true
+    # Check if the gmx executable exists
+    if isnothing(Sys.which(gmx))
+        throw(ArgumentError("""\n
+            Could not find GROMACS `$gmx` executable. Add the `gmx` command to the path or provide it explicitly with the `gmx` keyword argument.
+
+        """))
+    end
     index_file = tempname() * ".ndx"
     sasa_file = tempname() * ".xvg"
-    p = read_pdb(pdbname, "protein and not element H")
+    p = read_pdb(pdbname, at -> isprotein(at) && keepH(at))
     inds_protein = index.(p)
-    inds_sidechain = index.(select(p, "resname $resname and not (name N CA C O)"))
-    inds_backbone = index.(select(p, "resname $resname and name N CA C O"))
+    inds_sidechain = index.(select(p, at -> resname(at) == rname && sidechain(at)))
+    inds_backbone = index.(select(p, at -> resname(at) == rname && backbone(at)))
     open(index_file, "w") do io
         write(io, "[ SC ]\n")
         for i in inds_sidechain
@@ -245,18 +343,18 @@ function gmx_sasa_single(pdbname, resname)
     if length(inds_sidechain) == 0 # special case for GLY
         sasa_sc = 0.0
         try
-            run(pipeline(`gmx sasa -s $pdbname -probe 0.14 -ndots 500 -surface PROT -output BB -n $index_file -o $sasa_file`, stdout=devnull, stderr=devnull))
+            run(pipeline(`$gmx sasa -s $pdbname -probe 0.14 -ndots $n_dots -surface PROT -output BB -n $index_file -o $sasa_file`, stdout=devnull, stderr=devnull))
         catch
-            error("Error running gmx sasa for of $resname in $pdbname")
+            error("Error running $gmx sasa for of $resname in $pdbname")
         end
-        sasa_bb = read_gmx_sasa_values(sasa_file, 1)[1]
+        sasa_bb = read_gmx_delta_sasa_per_restype_values(sasa_file, 1)[1]
     else
         try
-            run(pipeline(`gmx sasa -s $pdbname -probe 0.14 -ndots 500 -surface PROT -output SC BB -n $index_file -o $sasa_file`, stdout=devnull, stderr=devnull))
+            run(pipeline(`$gmx sasa -s $pdbname -probe 0.14 -ndots $n_dots -surface PROT -output SC BB -n $index_file -o $sasa_file`, stdout=devnull, stderr=devnull))
         catch
-            error("Error running gmx sasa for of $resname in $pdbname")
+            error("Error running $gmx sasa for of $resname in $pdbname")
         end
-        sasa_temp = read_gmx_sasa_values(sasa_file, 2)
+        sasa_temp = read_gmx_delta_sasa_per_restype_values(sasa_file, 2)
         sasa_sc, sasa_bb = sasa_temp[1], sasa_temp[2]
     end
     rm(sasa_file, force=true)